->在帮助文档下方添加快捷按钮,新彩蛋 彩蛋:1-版本锁死999.999.999(无法更新) 2-isImportDay->false 3-CTColor->蓝,淡蓝 4-弹窗标题:骇客已入侵 5-按钮图标:error 6-无法关闭 7-软件信息修改 8-软件+弹窗图标修改 9-只能播放银狼PV 10-修改错误提示音 11-修改ETPanel文字

+++ src/com/wmp/classtools/infSet/panel/ClearTempPanel.java
@@ -2,6 +2,7 @@ package com.wmp.classTools.infSet.panel;
 
 import com.wmp.Main;
 import com.wmp.PublicTools.io.IOForInfo;
+import com.wmp.PublicTools.printLog.Log;
 import com.wmp.classTools.CTComponent.CTButton;
 import com.wmp.classTools.CTComponent.CTSetsPanel;
 
@@ -46,7 +47,7 @@ public class ClearTempPanel extends CTSetsPanel {
         this.add(buttonPanel);
     }
 
-    private static CTButton getButton(String name, String DATA_PATH) throws MalformedURLException {
+    private CTButton getButton(String name, String DATA_PATH) throws MalformedURLException {
         return new CTButton(CTButton.ButtonText, name,
                 "/image/%s/delete_0.png",
                 "/image/%s/delete_1.png", 35, 150, () -> {
@@ -54,7 +55,13 @@ public class ClearTempPanel extends CTSetsPanel {
         });
     }
 
-    private static void deleteDir(String DATA_PATH) {
+    private void deleteDir(String DATA_PATH) {
+
+        int i = Log.info.inputInt(this, "清理临时文件", "是否清理?");
+        if (i == JOptionPane.NO_OPTION) {
+            return;
+        }
+
         try {
             IOForInfo.deleteDirectoryRecursively(Paths.get(DATA_PATH));
